Remove num_QPUs, add tests.
